html/interaction/focus/processing-model/focus-fixup-rule-one-no-dialogs.html

<!DOCTYPE html>
<meta charset="utf-8">
<meta name="timeout" content="long">
<title>Focus fixup rule (no &lt;dialog&gt;s involved)</title>
<link rel="author" title="Domenic Denicola" href="mailto:d@domenic.me">
<link rel="help" href="https://html.spec.whatwg.org/multipage/interaction.html#focus-fixup-rule">
<link rel="help" href="https://html.spec.whatwg.org/multipage/forms.html#attr-fieldset-disabled">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div>
  <button id="button1">Button 1</button>
  <button id="button2">Button 2</button>
  <button id="button3">Button 3</button>
  <fieldset id="fieldset1"><button id="button4">Button 4</button></fieldset>
  <fieldset id="fieldset2" disabled><legend><button id="button5">Button 5</button></legend></fieldset>
  <div id="div1" tabindex="0">Div 1</div>
  <div id="editable1" contenteditable=true>Editor 1</div>
  <button id="button6">Button 6</button>
  <button id="button7">Button 7</button>
  <button id="button8">Button 8</button>
  <fieldset id="fieldset3"><button id="button9">Button 9</button></fieldset>
  <fieldset id="fieldset4" disabled><legend><button id="button10">Button 10</button></legend></fieldset>
  <div id="div2" tabindex="0">Div 2</div>
  <div id="editable2" contenteditable=true>Editor 2</div>
</div>

<script>
"use strict";

const runTestSync = function (t, el, test) {
  const tag = el.tagName.toLowerCase();
  el.focus();
  assert_equals(document.activeElement, el, `Sanity check: the ${tag} must start focused`);
  test();
  assert_not_equals(document.activeElement, el, `Immediately after test, the ${tag} must synchronously no longer be focused`);
  assert_equals(document.activeElement, document.body, "Immediately after test, the body must synchronously be focused");
}

const runTestAsync = function (t, el, test) {
  const tag = el.tagName.toLowerCase();
  el.focus();
  assert_equals(document.activeElement, el, `Sanity check: the ${tag} must start focused`);
  test();
  return t.step_wait(() => document.activeElement !== el, `After test, the ${tag} must no longer be focused`)
    .then(() => assert_equals(document.activeElement, document.body, "After test, the body must be focused"));
}

test(t => {
  const button = document.querySelector("#button1");
  runTestSync(t, button, () => button.disabled = true);
}, "Disabling the active element (making it expressly inert) [sync]");

promise_test(async t => {
  const button = document.querySelector("#button6");
  return await runTestAsync(t, button, () => button.disabled = true);
}, "Disabling the active element (making it expressly inert)");

test(t => {
  const button = document.querySelector("#button2");
  runTestSync(t, button, () => button.hidden = true);
}, "Hiding the active element [sync]");

promise_test(async t => {
  const button = document.querySelector("#button7");
  return await runTestAsync(t, button, () => button.hidden = true);
}, "Hiding the active element");

test(t => {
  const button = document.querySelector("#button3");
  runTestSync(t, button, () => button.remove());
}, "Removing the active element from the DOM [sync]");

promise_test(async t => {
  const button = document.querySelector("#button8");
  return await runTestAsync(t, button, () => button.remove());
}, "Removing the active element from the DOM");

test(t => {
  const fieldset = document.querySelector("#fieldset1");
  const button = document.querySelector("#button4");
  runTestSync(t, button, () => fieldset.disabled = true);
}, "Disabling <fieldset> affects its descendants [sync]");

promise_test(async t => {
  const fieldset = document.querySelector("#fieldset3");
  const button = document.querySelector("#button9");
  return await runTestAsync(t, button, () => fieldset.disabled = true);
}, "Disabling <fieldset> affects its descendants");

test(t => {
  const fieldset = document.querySelector("#fieldset2");
  const button = document.querySelector("#button5");
  const fn = () => fieldset.insertBefore(document.createElement("legend"), fieldset.firstChild);
  runTestSync(t, button, fn)
}, "Changing the first legend element in disabled <fieldset> [sync]");

promise_test(async t => {
  const fieldset = document.querySelector("#fieldset4");
  const button = document.querySelector("#button10");
  const fn = () => fieldset.insertBefore(document.createElement("legend"), fieldset.firstChild);
  return await runTestAsync(t, button, fn)
}, "Changing the first legend element in disabled <fieldset>");

test(t => {
  const div = document.querySelector("#div1");
  runTestSync(t, div, () => div.removeAttribute("tabindex"));
}, "Removing the tabindex attribute from a div [sync]");

promise_test(async t => {
  const div = document.querySelector("#div2");
  return await runTestAsync(t, div, () => div.removeAttribute("tabindex"));
}, "Removing the tabindex attribute from a div");

test(t => {
  const div = document.querySelector("#editable1");
  runTestSync(t, div, () => div.contentEditable = false);
}, "Disabling contenteditable [sync]");

promise_test(async t => {
  const div = document.querySelector("#editable2");
  return await runTestAsync(t, div, () => div.contentEditable = false);
}, "Disabling contenteditable");
</script>
